Revbump sys-boot/rpi3-boot-config to r3 (retain fkms)
In the end, retaining fkms for 1.4.0 release on both pi-top and
regular variants.

For pi-top, because it is necessary to allow audio output on the pi-top
speaker.
For regular, because it plays nicely with display manager compositing,
which true kms currently does not.
Also, with compositing off, there is only about a 5% bandwidth
improvement for gl through the use of true kms.