Add pulseaudio-12.2-r3.ebuild (with rpi-deglitch USE)
Per e.g.
https://forum.manjaro.org/t/manjaro-arm-19-08-released/99031/72
pulseaudio on the RPi4 (at least) can be improved (for certain
use cases) by setting the tsched=0 parameter to module-udev-detect
(in /etc/pulse/default.pa).

This commit just adds a tweaked pulseaudio-12.2-r3.ebuild to the
overlay which, when emerged with the rpi-deglitch USE flag set,
applies this fix to the default.pa file upon installation.