profiles: don't use wildcard in accept_keywords for packages keyworded for arm64
Most of entries in profiles/targets/rpi3/package.accept_keywords/*
specify value "* ~*", which means "package is visible if it is in stable
or testing status on any architecture".

This means, once any affected package gets a new ebuild having *any*
keyword, e.g. ~amd64, it becomes "visible" as a pending update for the
system. This is not a bad trick per se, especially for a deliberate
"bleeding edge" system, and especially in earlier times when arm64
didn't maintain a stable keyword. But this means potentially dealing
with more cases of breakage, because Gentoo staff hasn't yet assured
that the new version works on arm64 architecture.

This commit turns such lines, for packages which are keyworded for
arm64 (at least in unstable status, ~arm64), into lines with just the
package name, which means "package is visible if it is in stable or
testing status on *host* architecture". This means arm64 in this case.
It might lead to some confusion for people who are trying to use the
overlay for arm (32-bit) systems, but they don't seem to be the target
audience anyway. I am open to changing this commit to list arm64
explicitly.

The script used to produce the change:

    REPO=/usr/portage
    cat profiles/targets/rpi3/package.accept_keywords/* \
    | grep -v -E '^$|^#|^=' \
    | awk '{ print $1 }' \
    | sort -u \
    | while read CATPN; do
          if grep -q 'KEYWORDS=.*arm64' $REPO/$CATPN/*.ebuild &>/dev/null; then
    	  # Keyworded, stable or unstable
    	  # Replace '$CAT/$PN * ~*' which is 'accept stable or testing on any arch'
    	  # with '$CAT/$PN' which is 'accept testing on host arch'
    	  sed -i -e "s:^${CATPN}[ ]\+[*] [~][*]$:${CATPN}:" profiles/targets/rpi3/package.accept_keywords/*
          fi
      done