Don't set USE=-sanitize in the profile. Causes GCC to be rebuilt (#119)
Because USE=-sanitize is set in the genpi64 profile, we end up building GCC again during the build process. The default settings from upstream have USE=sanitize, not USE=-sanitize.

GCC is one of the most time consuming packages to build, so we're better off just keeping the upstream settings to avoid rebuilds.